+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Release
 
+- interquartile_range function was added (24/05/2021)
 - Dixon class was moved to outliers module (22/06/2022)
 - Added the ModifiedZScore Class (22/06/2022)
 - Added the ZScore Class (22/06/2022)